# backend/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Coworking API",
    "description": "Автоматически сгенерированная документация",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/api/workstations/": {
      "get": {
        "summary": "Получить все рабочие места (с фильтрацией по этажу)",
        "description": "",
        "parameters": [
          {
            "name": "floorId",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "Создать рабочее место",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "number": {
                  "example": "any"
                },
                "floorId": {
                  "example": "any"
                },
                "type": {
                  "example": "any"
                },
                "capacity": {
                  "example": "any"
                },
                "basePricePerHour": {
                  "example": "any"
                },
                "basePricePerDay": {
                  "example": "any"
                },
                "basePricePerWeek": {
                  "example": "any"
                },
                "basePricePerMonth": {
                  "example": "any"
                },
                "x": {
                  "example": "any"
                },
                "y": {
                  "example": "any"
                },
                "width": {
                  "example": "any"
                },
                "height": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/workstations/{id}": {
      "get": {
        "summary": "Получить конкретное рабочее место",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "summary": "Обновить рабочее место",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "number": {
                  "example": "any"
                },
                "type": {
                  "example": "any"
                },
                "capacity": {
                  "example": "any"
                },
                "basePricePerHour": {
                  "example": "any"
                },
                "basePricePerDay": {
                  "example": "any"
                },
                "basePricePerWeek": {
                  "example": "any"
                },
                "basePricePerMonth": {
                  "example": "any"
                },
                "x": {
                  "example": "any"
                },
                "y": {
                  "example": "any"
                },
                "width": {
                  "example": "any"
                },
                "height": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "summary": "Удалить рабочее место",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/workstations/floor/{floorId}": {
      "get": {
        "summary": "Получить все рабочие места на этаже",
        "description": "",
        "parameters": [
          {
            "name": "floorId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/workstations/availability/{id}": {
      "get": {
        "summary": "Проверить доступность рабочего места",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "date",
            "in": "query",
            "type": "string"
          },
          {
            "name": "time",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/workstations/{id}/inventory": {
      "patch": {
        "summary": "Обновить инвентарь рабочего места",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "inventory": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/inventory-items/workstation/{workstationId}": {
      "get": {
        "summary": "Получить инвентарь для конкретного рабочего места (комнаты)",
        "description": "",
        "parameters": [
          {
            "name": "workstationId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/inventory-items/{id}/workstation": {
      "patch": {
        "summary": "Привязывает инвентарь к рабочему месту",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "workstationId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/bookings/coworking/{coworkingCenterId}/workstation/{workstationId}": {
      "get": {
        "summary": "Получить бронирования конкретного рабочего места",
        "description": "",
        "parameters": [
          {
            "name": "coworkingCenterId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "workstationId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "startDate",
            "in": "query",
            "type": "string"
          },
          {
            "name": "endDate",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}